==> 3_FinalReport.qmd <==
---
title: "Fractures in SUS"
author: "Bruno Casaes Teixeira"
format:
  html:
    code-fold: true
    code-summary: "Show the code"
    code-tools: true
    embed-resources: true
    df-print: paged
    page-layout: full
editor: visual
number-sections: true
execute:
  warning: false
  freeze: auto
toc: true
fig-width: 8
fig-asp: 0.618
---

```{r}
#| label: raw-data
#| echo: false


library(tidyverse)
library(forecast)
library(scales)
library(epitools)
library(downloadthis)
source("2_DataProcessing.R")

```

# Methods:

**Database:**

The population data for Brazil was obtained from the Brazilian Ministry of Health's website \[http://tabnet.datasus.gov.br/cgi/deftohtm.exe?ibge/cnv/popsvsbr.def\]. This publicly accessible source offers a detailed overview of the country's demographic information.

Separately, data pertaining to the national private health insurance was sourced from the website of the National Supplementary Health Agency (ANS) \[https://www.ans.gov.br/anstabnet/cgi-bin/dh?dados/tabnet_02.def\], the body responsible for the regulation and supervision of the private health insurance sector in Brazil.

Data related to densitometries and hospitalizations were retrieved from two distinct platforms maintained by the Brazilian Ministry of Health. The Ambulatory Information System (SIA-SUS) provided the densitometry data, while the Hospital Information System (SIH-SUS) supplied data on hospitalizations \[https://datasus.saude.gov.br/transferencia-de-arquivos/\].

**Data Processing:**

Upon retrieval, the data from both SIA-SUS and SIH-SUS were processed to isolate cases of fractures. This was accomplished by applying a regular expression (regex) pattern to the ICD-10 codes in the datasets. The pattern used was `^S.{1}2.{0,1}$`, designed to match specific codes corresponding to fracture hospitalizations.

In constructing the denominator for the subsequent analysis, we considered the total population and subtracted the population covered by private health insurance. This assumption is based on the premise that individuals with private insurance would not utilize the public healthcare system for their healthcare needs.

**Statirtical Analysis**

For the prediction of hip fractures over the next seven years, we employed the auto.arima function from the forecast package in R. This function uses the auto-regressive integrated moving average (ARIMA) model to predict future data points on a time series. The predicted number of hip fractures was then compared with the actual recorded figures.

**Code lists**

ICD codes used to filter (subcodes where also used).

```{r}
#| label: icd-codes
CID %>% 
  filter(grepl(CD_COD, pattern = "^S.{1}2.{0,1}$") & CID == CID_GRUPO) %>% 
  select('ICD Code' =  CD_COD, 'ICD Group' = CID_GRUPO)
```

Specific to hip fracture, the codes used were:

```{r}
#| label: icd-codes-fracture
CID %>% 
  filter(CD_COD %in% c("S720","S721","S722")) %>% 
  select('ICD Code' =  CD_COD, 'ICD Description' = CID)

```

To identify procedures that would uniquely identify a fracture, we have used the following list of procedures codes:

```{r}
#| label: unique_procedures
FiltraProc %>% 
  mutate(IP_COD = as.numeric(Procedimentos)) %>% 
  left_join(tbProc %>% mutate(IP_COD = as.numeric(IP_COD)), by = "IP_COD") %>% 
  select('Procedure Code' = Procedimentos, 'Procedure Description'= PROC_DESC)
```

# Results

## General Fractures and Age

::: panel-tabset
### Total

```{r}
#| label: total-fracture
#|fig-width: 10
FracTotRate <- RD %>% filter (year(DT_INTER) == 2022) %>% 
  group_by(FaixaEtr, DIAG_CAP, DS_CID_CAP) %>%
  summarize(n = n_distinct(N_AIH), .groups = "keep") %>% 
  left_join(popSUSBR %>% filter(ano == 2022) %>% group_by(FaixaEtr) %>% summarise(PopSUS = sum(PopSUS)), by = "FaixaEtr") %>%
  mutate(IndFrac = n*100000/PopSUS)

FracTotRate72 <- RD %>% filter (year(DT_INTER) == 2022) %>%  filter (DIAG_CAP =="S72")%>%
  group_by(FaixaEtr, DS_CID, DIAG_CAP) %>% summarize(n = n_distinct(N_AIH), .groups = "keep") %>% 
  left_join(popSUSBR %>% filter(ano == 2022) %>% group_by(FaixaEtr) %>% summarise(PopSUS = sum(PopSUS)), by = "FaixaEtr") %>%
  mutate(IndFrac = n*100000/PopSUS)

levels(FracTotRate$FaixaEtr) <- c("0 to 4 years", "5 to 9 years", "10 to 14 years", "15 to 19 years", "20 to 24 years", "25 to 29 years", "30 to 34 years", "35 to 39 years", "40 to 44 years", "45 to 49 years", "50 to 54 years", "55 to 59 years", "60 to 64 years", "65 to 69 years", "70 to 74 years", "75 to 79 years", "80 years or more")

levels(FracTotRate72$FaixaEtr) <- c("0 to 4 years", "5 to 9 years", "10 to 14 years", "15 to 19 years", "20 to 24 years", "25 to 29 years", "30 to 34 years", "35 to 39 years", "40 to 44 years", "45 to 49 years", "50 to 54 years", "55 to 59 years", "60 to 64 years", "65 to 69 years", "70 to 74 years", "75 to 79 years", "80 years or more")

#Hospitalizations by Fracture Type and Age Group (2017) (PORTUGUES)
FracLabel <- (FracTotRate %>% arrange (desc(n)))[c(1,5,7),]
FracLabel$DS_CID_CAP_en <- c("S72 Fracture of femur", "S82 Fracture of lower leg", "S52 Fracture of forearm")


p <- ggplot (FracTotRate, aes(x = FaixaEtr, y=n, group = DIAG_CAP,
                         color = case_when(DIAG_CAP== "S72" ~ "green",
                                           DIAG_CAP== "S52" ~ "blue",
                                           DIAG_CAP== "S82" ~ "red",
                                           FALSE ~ "black")
                         )) +
    geom_line(aes(linewidth = replace_na(case_when(DIAG_CAP== "S72" ~ 1.2,
                                              DIAG_CAP== "S52" ~ 1.2,
                                              DIAG_CAP== "S82" ~ 1.2,
                                              FALSE ~ 1),1)))+
    geom_text(aes(label = DS_CID_CAP_en), data = FracLabel, nudge_y = 700, nudge_x = c(-2,0,2), color = "black") +
    labs(title = "Hospitalizations by type of Fracture and Age Group (2022)", x = "Age Group", y = "Hospitalizations", caption = "Source: SIH-DATASUS") +
      theme_bw() +
    theme(axis.text.x = element_text(angle = 20, hjust = 1))+
    scale_colour_discrete(guide = FALSE) +
     scale_linewidth_continuous(guide = "none", limits = c(1,10))

p

download_this(p, output_extension = "svg")
```

The three most frequent fracture types for hospitalizations are Femur, Leg including ankle, and forearm. They significantly vary across age groups. But clearly, the most frequent fracture causing hospitalizations, even in absolute number, is femoral fractures in the elderly.

```{r}
#| label: total-fracture-detailed
FracTotRate72$DS_CID_en <- factor(FracTotRate72$DS_CID)


levels(FracTotRate72$DS_CID_en) <- c("S72 Fracture of femur", "S72.0 Fracture of neck of femur", "S72.1 Pertrochanteric fracture", "S72.2 Subtrochanteric fracture", "S72.3 Fracture of shaft of femur", "S72.4 Fracture of lower end of femur", "S72.7 Multiple fractures of femur", "S72.8 Fractures of other parts of femur", "S72.9 Fracture of femur, part unspecified")

p <- p + geom_area(data = FracTotRate72, aes(x = FaixaEtr, y=n, group = DS_CID_en, fill = DS_CID_en), position = "stack", stat = "identity", alpha = 0.5) +labs(fill = "Femoral Fracture Type")

p
download_this(p, output_extension = "svg")

```

In the plot above we see the different types of femoral fracture, being the most common around 20-24 years old is the shaft of the femur, while hip fracture is the most common among the elderly.

```{r}
#| label: total-fracture-rate
#Hospitalizations RATE by Fracture Type and Age Group (2022) (PORTUGUES)
FracLabelInd <- FracTotRate %>% arrange (desc(IndFrac)) %>% group_by (DIAG_CAP) %>% filter(DIAG_CAP %in% c("S72", "S52", "S82")) %>% slice (1) 

FracLabelInd$DS_CID_CAP_en <- c("S52 Fracture of forearm", "S72 Fracture of femur", "S82 Fracture of lower leg")

p <- ggplot (FracTotRate, aes(x = FaixaEtr, y=IndFrac, group = DIAG_CAP,
                         color = case_when(DIAG_CAP== "S72" ~ "green",
                                           DIAG_CAP== "S52" ~ "blue",
                                           DIAG_CAP== "S82" ~ "red",
                                           FALSE ~ "black")
)) +
  geom_line(aes(linewidth = replace_na(case_when(DIAG_CAP== "S72" ~ 1.2,
                                            DIAG_CAP== "S52" ~ 1.2,
                                            DIAG_CAP== "S82" ~ 1.2,
                                            FALSE ~ 1),1)))+
  geom_text(aes(label = DS_CID_CAP_en), data = FracLabelInd, nudge_x= c(0,-2,3), nudge_y = c(50, 30, 50), color = "black") +
  labs(title = "Hospitalization Rate (per 100,000 inhabitants) by type of Fracture and Age Group (2022)", x = "Age Group", y = "Hospitalization Rate (per 100,000 inhabitants)", caption = "Source: SIH -DATASUS, IBGE, ANS") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 20, hjust = 1))+
  scale_colour_discrete(guide = FALSE) +
  scale_linewidth_continuous(guide = FALSE, limits = c(1,10))

p

download_this(p, output_extension = "svg")
```

Previously we saw in absolute numbers, but considering that above 80 years old population is much less numerous in comparison to the 20s, after we adjust hospitalization rates over population, we observe that the risk is substantially higher in the elderly.

```{r}
#| label: total-fracture-rate-detailed
#|fig-width: 10

FracTotRate72$DS_CID_en <- factor(FracTotRate72$DS_CID)

levels(FracTotRate72$DS_CID_en) <- c("S72 Fracture of femur", "S72.0 Fracture of neck of femur", "S72.1 Pertrochanteric fracture", "S72.2 Subtrochanteric fracture", "S72.3 Fracture of shaft of femur", "S72.4 Fracture of lower end of femur", "S72.7 Multiple fractures of femur", "S72.8 Fractures of other parts of femur", "S72.9 Fracture of femur, part unspecified")

p <- p + geom_area(data = FracTotRate72, aes(x = FaixaEtr, y=IndFrac, group = DS_CID_en, fill = DS_CID_en), position = "stack", stat = "identity", alpha = 0.5) +labs(fill = "Femoral Fracture Type")

p
download_this(p, output_extension = "svg")
```

The same plot, now considering the different types of femoral fracture.

**Observation:** The plots above consider all hospitalizations, including re-hospitalization for the same surgery.

### Female

```{r}
#| label: female-fracture
#|fig-width: 10
FracTotRate <- RD %>% filter (year(DT_INTER) == 2022, SEXO == "Female") %>% 
  group_by(FaixaEtr, DIAG_CAP, DS_CID_CAP) %>%
  summarize(n = n_distinct(N_AIH), .groups = "keep") %>% 
  left_join(popSUSBR %>% filter(ano == 2022) %>% group_by(FaixaEtr) %>% summarise(PopSUS = sum(PopSUS)), by = "FaixaEtr") %>%
  mutate(IndFrac = n*100000/PopSUS)

FracTotRate72 <- RD %>% filter (year(DT_INTER) == 2022, SEXO == "Female") %>%  filter (DIAG_CAP =="S72")%>%
  group_by(FaixaEtr, DS_CID, DIAG_CAP) %>% summarize(n = n_distinct(N_AIH), .groups = "keep") %>% 
  left_join(popSUSBR %>% filter(ano == 2022) %>% group_by(FaixaEtr) %>% summarise(PopSUS = sum(PopSUS)), by = "FaixaEtr") %>%
  mutate(IndFrac = n*100000/PopSUS)

levels(FracTotRate$FaixaEtr) <- c("0 to 4 years", "5 to 9 years", "10 to 14 years", "15 to 19 years", "20 to 24 years", "25 to 29 years", "30 to 34 years", "35 to 39 years", "40 to 44 years", "45 to 49 years", "50 to 54 years", "55 to 59 years", "60 to 64 years", "65 to 69 years", "70 to 74 years", "75 to 79 years", "80 years or more")

levels(FracTotRate72$FaixaEtr) <- c("0 to 4 years", "5 to 9 years", "10 to 14 years", "15 to 19 years", "20 to 24 years", "25 to 29 years", "30 to 34 years", "35 to 39 years", "40 to 44 years", "45 to 49 years", "50 to 54 years", "55 to 59 years", "60 to 64 years", "65 to 69 years", "70 to 74 years", "75 to 79 years", "80 years or more")

FracTotRate72$DS_CID_en <- factor(FracTotRate72$DS_CID)


levels(FracTotRate72$DS_CID_en) <- c("S72 Fracture of femur", "S72.0 Fracture of neck of femur", "S72.1 Pertrochanteric fracture", "S72.2 Subtrochanteric fracture", "S72.3 Fracture of shaft of femur", "S72.4 Fracture of lower end of femur", "S72.7 Multiple fractures of femur", "S72.8 Fractures of other parts of femur", "S72.9 Fracture of femur, part unspecified")

#Hospitalizations by Fracture Type and Age Group (2017) (PORTUGUES)
FracLabel <- (FracTotRate %>% arrange (desc(n)))[c(1,5,18),]
FracLabel$DS_CID_CAP_en <- c("S72 Fracture of femur", "S52 Fracture of forearm", "S82 Fracture of lower leg" )

p <- ggplot (FracTotRate, aes(x = FaixaEtr, y=n, group = DIAG_CAP,
                         color = case_when(DIAG_CAP== "S72" ~ "green",
                                           DIAG_CAP== "S52" ~ "blue",
                                           DIAG_CAP== "S82" ~ "red",
                                           FALSE ~ "black")
                         )) +
    geom_line(aes(linewidth = replace_na(case_when(DIAG_CAP== "S72" ~ 1.2,
                                              DIAG_CAP== "S52" ~ 1.2,
                                              DIAG_CAP== "S82" ~ 1.2,
                                              FALSE ~ 1),1)))+
    geom_text(aes(label = DS_CID_CAP_en), data = FracLabel, nudge_y = c(700,700,2000), nudge_x = c(-2,0,2), color = "black") +
    labs(title = "Hospitalizations by type of Fracture and Age Group (2022)", x = "Age Group", y = "Hospitalizations", caption = "Source: SIH-DATASUS") +
    theme_bw() +
    theme(axis.text.x = element_text(angle = 20, hjust = 1))+
    scale_colour_discrete(guide = FALSE) +
     scale_linewidth_continuous(guide = "none", limits = c(1,10))

p
download_this(p, output_extension = "svg")
```

The three most frequent fracture types for hospitalizations are Femur, Leg including ankle, and forearm. They significantly vary across age groups. But clearly, the most frequent fracture causing hospitalizations, even in absolute number, is femoral fractures in the elderly.

```{r}
#| label: female-fracture-detailed
p <- p + geom_area(data = FracTotRate72, aes(x = FaixaEtr, y=n, group = DS_CID_en, fill = DS_CID_en), position = "stack", stat = "identity", alpha = 0.5) +labs(fill = "Femoral Fracture Type")

p
  
download_this(p, output_extension = "svg")
```

In the plot above we see the different types of femoral fracture, being the most common around 20-24 years old is the shaft of the femur, while hip fracture is the most common among the elderly.

```{r}
#| label: female-fracture-rate
#Hospitalizations RATE by Fracture Type and Age Group (2022) (PORTUGUES)
FracLabelInd <- FracTotRate %>% arrange (desc(IndFrac)) %>% group_by (DIAG_CAP) %>% filter(DIAG_CAP %in% c("S72", "S52", "S82")) 

FracLabelInd <- FracLabelInd[c(1,18,35),]

FracLabelInd$DS_CID_CAP_en <- c( "S72 Fracture of femur","S82 Fracture of lower leg", "S52 Fracture of forearm" )




p <- ggplot (FracTotRate, aes(x = FaixaEtr, y=IndFrac, group = DIAG_CAP,
                         color = case_when(DIAG_CAP== "S72" ~ "green",
                                           DIAG_CAP== "S52" ~ "blue",
                                           DIAG_CAP== "S82" ~ "red",
                                           FALSE ~ "black")
)) +
  geom_line(aes(linewidth = replace_na(case_when(DIAG_CAP== "S72" ~ 1.2,
                                            DIAG_CAP== "S52" ~ 1.2,
                                            DIAG_CAP== "S82" ~ 1.2,
                                            FALSE ~ 1),1)))+
  geom_text(aes(label = DS_CID_CAP_en), data = FracLabelInd, nudge_x= c(-2,-1,1), nudge_y = c(50, 70, 50), color = "black") +
  labs(title = "Hospitalization Rate (per 100,000 inhabitants) by type of Fracture and Age Group (2022)", x = "Age Group", y = "Hospitalization Rate (per 100,000 inhabitants)", caption = "Source: SIH -DATASUS, IBGE, ANS") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 20, hjust = 1))+
  scale_colour_discrete(guide = FALSE) +
  scale_linewidth_continuous(guide = FALSE, limits = c(1,10))

p
download_this(p, output_extension = "svg")
```

Previously we saw in absolute numbers, but considering that above 80 years old population is much less numerous in comparison to the 20s, after we adjust hospitalization rates over population, we observe that the risk is substantially higher in the elderly.

```{r}
#| label: female-fracture-rate-detailed
#|fig-width: 10
p <- p + geom_area(data = FracTotRate72, aes(x = FaixaEtr, y=IndFrac, group = DS_CID_en, fill = DS_CID_en), position = "stack", stat = "identity", alpha = 0.5) +labs(fill = "Femoral Fracture Type")

p
download_this(p, output_extension = "svg")
```

The same plot, now considering the different types of femoral fracture.

**Observation:** The plots above consider all hospitalizations, including re-hospitalization for the same surgery.

### Male

```{r}
#| label: male-fracture
#|fig-width: 10
FracTotRate <- RD %>% filter (year(DT_INTER) == 2022, SEXO == "Male") %>% 
  group_by(FaixaEtr, DIAG_CAP, DS_CID_CAP) %>%
  summarize(n = n_distinct(N_AIH), .groups = "keep") %>% 
  left_join(popSUSBR %>% filter(ano == 2022) %>% group_by(FaixaEtr) %>% summarise(PopSUS = sum(PopSUS)), by = "FaixaEtr") %>%
  mutate(IndFrac = n*100000/PopSUS)

FracTotRate72 <- RD %>% filter (year(DT_INTER) == 2022, SEXO == "Male") %>%  filter (DIAG_CAP =="S72")%>%
  group_by(FaixaEtr, DS_CID, DIAG_CAP) %>% summarize(n = n_distinct(N_AIH), .groups = "keep") %>% 
  left_join(popSUSBR %>% filter(ano == 2022) %>% group_by(FaixaEtr) %>% summarise(PopSUS = sum(PopSUS)), by = "FaixaEtr") %>%
  mutate(IndFrac = n*100000/PopSUS)

levels(FracTotRate$FaixaEtr) <- c("0 to 4 years", "5 to 9 years", "10 to 14 years", "15 to 19 years", "20 to 24 years", "25 to 29 years", "30 to 34 years", "35 to 39 years", "40 to 44 years", "45 to 49 years", "50 to 54 years", "55 to 59 years", "60 to 64 years", "65 to 69 years", "70 to 74 years", "75 to 79 years", "80 years or more")

levels(FracTotRate72$FaixaEtr) <- c("0 to 4 years", "5 to 9 years", "10 to 14 years", "15 to 19 years", "20 to 24 years", "25 to 29 years", "30 to 34 years", "35 to 39 years", "40 to 44 years", "45 to 49 years", "50 to 54 years", "55 to 59 years", "60 to 64 years", "65 to 69 years", "70 to 74 years", "75 to 79 years", "80 years or more")

FracTotRate72$DS_CID_en <- factor(FracTotRate72$DS_CID)


levels(FracTotRate72$DS_CID_en) <- c("S72 Fracture of femur", "S72.0 Fracture of neck of femur", "S72.1 Pertrochanteric fracture", "S72.2 Subtrochanteric fracture", "S72.3 Fracture of shaft of femur", "S72.4 Fracture of lower end of femur", "S72.7 Multiple fractures of femur", "S72.8 Fractures of other parts of femur", "S72.9 Fracture of femur, part unspecified")


#Hospitalizations by Fracture Type and Age Group (2017) (PORTUGUES)
FracLabel <- (FracTotRate %>% arrange (desc(n)))[c(1,5,8),]
FracLabel$DS_CID_CAP_en <- c("S52 Fracture of forearm",  "S82 Fracture of lower leg","S72 Fracture of femur" )


p <- ggplot (FracTotRate, aes(x = FaixaEtr, y=n, group = DIAG_CAP,
                         color = case_when(DIAG_CAP== "S72" ~ "green",
                                           DIAG_CAP== "S52" ~ "blue",
                                           DIAG_CAP== "S82" ~ "red",
                                           FALSE ~ "black")
                         )) +
    geom_line(aes(linewidth = replace_na(case_when(DIAG_CAP== "S72" ~ 1.2,
                                              DIAG_CAP== "S52" ~ 1.2,
                                              DIAG_CAP== "S82" ~ 1.2,
                                              FALSE ~ 1),1)))+
    geom_text(aes(label = DS_CID_CAP_en), data = FracLabel, nudge_y = 700, nudge_x = c(0,0,-2), color = "black") +
    labs(title = "Hospitalizations by type of Fracture and Age Group (2022)", x = "Age Group", y = "Hospitalizations", caption = "Source: SIH-DATASUS") +
    theme_bw() +
    theme(axis.text.x = element_text(angle = 20, hjust = 1))+
    scale_colour_discrete(guide = FALSE) +
     scale_linewidth_continuous(guide = "none", limits = c(1,10))

p
download_this(p, output_extension = "svg")
```

The three most frequent fracture types for hospitalizations are Femur, Leg including ankle, and forearm. They significantly vary across age groups. But clearly, the most frequent fracture causing hospitalizations, even in absolute number, is femoral fractures in the elderly.

```{r}
#| label: male-fracture-detailed
p <- p + geom_area(data = FracTotRate72, aes(x = FaixaEtr, y=n, group = DS_CID_en, fill = DS_CID_en), position = "stack", stat = "identity", alpha = 0.5) +labs(fill = "Femoral Fracture Type")
  
p
download_this(p, output_extension = "svg")
```

In the plot above we see the different types of femoral fracture, being the most common around 20-24 years old is the shaft of the femur, while hip fracture is the most common among the elderly.

```{r}
#| label: male-fracture-rate
#Hospitalizations RATE by Fracture Type and Age Group (2022) (PORTUGUES)
FracLabelInd <- FracTotRate %>% arrange (desc(IndFrac)) %>% group_by (DIAG_CAP) %>% filter(DIAG_CAP %in% c("S72", "S52", "S82")) %>% slice (1) 

FracLabelInd$DS_CID_CAP_en <- c( "S52 Fracture of forearm","S72 Fracture of femur", "S82 Fracture of lower leg" )


p <- ggplot (FracTotRate, aes(x = FaixaEtr, y=IndFrac, group = DIAG_CAP,
                         color = case_when(DIAG_CAP== "S72" ~ "green",
                                           DIAG_CAP== "S52" ~ "blue",
                                           DIAG_CAP== "S82" ~ "red",
                                           FALSE ~ "black")
)) +
  geom_line(aes(linewidth = replace_na(case_when(DIAG_CAP== "S72" ~ 1.2,
                                            DIAG_CAP== "S52" ~ 1.2,
                                            DIAG_CAP== "S82" ~ 1.2,
                                            FALSE ~ 1),1)))+
  geom_text(aes(label = DS_CID_CAP), data = FracLabelInd, nudge_x= c(0,-2,3), nudge_y = c(20, 10, 20), color = "black") +
  labs(title = "Hospitalization Rate (per 100,000 inhabitants) by type of Fracture and Age Group (2022)", x = "Age Group", y = "Hospitalization Rate (per 100,000 inhabitants)", caption = "Source: SIH -DATASUS, IBGE, ANS") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 20, hjust = 1))+
  scale_colour_discrete(guide = FALSE) +
  scale_linewidth_continuous(guide = FALSE, limits = c(1,10))

p
download_this(p, output_extension = "svg")
```

Previously we saw in absolute numbers, but considering that above 80 years old population is much less numerous in comparison to the 20s, after we adjust hospitalization rates over population, we observe that the risk is substantially higher in the elderly.

```{r}
#| label: male-fracture-rate-detailed
#|fig-width: 10
p <- p + geom_area(data = FracTotRate72, aes(x = FaixaEtr, y=IndFrac, group = DS_CID_en, fill = DS_CID_en), position = "stack", stat = "identity", alpha = 0.5) +labs(fill = "Femoral Fracture Type")

p
download_this(p, output_extension = "svg")
```

The same plot, now considering the different types of femoral fracture.

**Observation:** The plots above consider all hospitalizations, including re-hospitalization for the same surgery.
:::

## Hip Fracture and Densitometries

```{r}
#| label: densitometries-forecast
densano <- PA %>% 
  group_by(ano) %>% 
  summarise(nDens = sum(PA_QTDPRO), .groups = "keep")


model <- auto.arima(densano$nDens[1:7])
forecast <- forecast(model)
forecast<- as.data.frame(forecast)

densano$nDensPred <- NA
densano$nDensPredlo80<- NA
densano$nDensPredlo95<- NA
densano$nDensPredhi80<- NA
densano$nDensPredhi95<- NA


densano$nDensPred[7] <- densano$nDens[7]
densano$nDensPredlo80[7] <- densano$nDens[7]
densano$nDensPredlo95[7] <- densano$nDens[7]
densano$nDensPredhi80[7] <- densano$nDens[7]
densano$nDensPredhi95[7] <- densano$nDens[7]
densano$nDensPred[8:15] <- forecast$`Point Forecast`[1:8]
densano$nDensPredlo80[8:15] <- forecast$`Lo 80`[1:8]
densano$nDensPredlo95[8:15] <- forecast$`Lo 95`[1:8]
densano$nDensPredhi80[8:15] <- forecast$`Hi 80`[1:8]
densano$nDensPredhi95[8:15] <- forecast$`Hi 95`[1:8]



p <- densano%>% 
  ggplot(aes(x = ano, y = nDens)) +
        geom_line(linewidth = 1) +
        geom_point(size = 3)+
        geom_ribbon(aes(ymin = nDensPredlo95, ymax = nDensPredhi95), alpha = 0.5, fill = "darkgray") +
        geom_ribbon(aes(ymin = nDensPredlo80, ymax = nDensPredhi80), alpha = 1, fill = "darkgray") +
        geom_line(aes(y = nDensPred), color = "blue") +
        geom_line(linewidth = 1) +
        geom_point(size = 3)+
        theme_bw() + 
  #y axis to be shown in not in scientific notation
        scale_y_continuous(labels = comma) +
  
        #scale_y_continuous(limits = c(0,600000), labels = comma) +
        labs(x = "Ano", y = "Densitometries Number", title = "Densitometries in Brazil (2008-2022)",
              subtitle = "Compared to an ARIMA model using data from 2008 to 2014", caption = "Source: SIH-DATASUS")
p
download_this(p, output_extension = "svg")
```

The provided plot traces the trajectory of densitometry examinations in Brazil from 2008 to 2022. From 2008 to 2014, we witness a robust growth in the number of densitometry exams, with the count more than doubling over this six-year period. This trend, however, seems to taper off around 2014, and in the subsequent eight years leading up to 2022, the growth rate of densitometry exams slowed considerably to around 10%.

Notably, the impact of the COVID-19 pandemic is clearly visible in the significant decline in densitometry exams conducted in 2020. Compared to the preceding year, 2019, the number of exams fell by 32%—a stark testament to the disruption caused by the global health crisis.

Encouragingly, the data from 2022 suggests a recovery from this downturn, with the number of densitometry exams appearing to return to pre-pandemic levels. However, the long-term effects of the pandemic-induced interruption in routine health screenings, such as densitometry tests, remain to be fully understood.

Table for the plot above:

```{r}
#| label: densitometries-forecast-data
PA %>% 
  group_by(ano) %>% 
  summarise(nDens = sum(PA_QTDPRO), .groups = "keep")


densano %>% 
  mutate('80% Confience Interval' = if_else(is.na(nDensPred), NA_character_, paste(round(nDensPredlo80,0), " - ", round(nDensPredhi80,0))),
         '95% Confience Interval' = if_else(is.na(nDensPred), NA_character_, paste(round(nDensPredlo95,0), " - ", round(nDensPredhi95,0)))) %>% 
    rename(Ano = ano, 
         'Number of Densitometries' = nDens,
         'Predicted Number of Densitometries' = nDensPred) %>% 
  select(-nDensPredlo80,-nDensPredhi80,-nDensPredlo95,-nDensPredhi95)
```

```{r}
#| label: densitometries-forecast-year

fracano <- RD %>% 
  filter(UNICO & CD_CID %in% c("S720","S721","S722")) %>% 
  filter(IDADE >= 50) %>% 
  filter(ano >= 2008) %>% 
  group_by(ano) %>% 
  summarise(nFrac = n_distinct(N_AIH))


model <- auto.arima(fracano$nFrac[1:8])
forecast <- forecast(model)
forecast<- as.data.frame(forecast)

fracano$nFracPred <- NA
fracano$nFracPredlo80<- NA
fracano$nFracPredlo95<- NA
fracano$nFracPredhi80<- NA
fracano$nFracPredhi95<- NA


fracano$nFracPred[8] <- fracano$nFrac[8]
fracano$nFracPredlo80[8] <- fracano$nFrac[8]
fracano$nFracPredlo95[8] <- fracano$nFrac[8]
fracano$nFracPredhi80[8] <- fracano$nFrac[8]
fracano$nFracPredhi95[8] <- fracano$nFrac[8]
fracano$nFracPred[9:15] <- forecast$`Point Forecast`[1:7]
fracano$nFracPredlo80[9:15] <- forecast$`Lo 80`[1:7]
fracano$nFracPredlo95[9:15] <- forecast$`Lo 95`[1:7]
fracano$nFracPredhi80[9:15] <- forecast$`Hi 80`[1:7]
fracano$nFracPredhi95[9:15] <- forecast$`Hi 95`[1:7]

p<- fracano %>% 
  ggplot(aes(x = ano, y = nFrac)) +
        theme_bw() +
        scale_y_continuous(limits = c(0,70000), labels = comma) +
        labs(x = "Year", y = "Number of Fractures") +
        geom_ribbon(aes(ymin = nFracPredlo95, ymax = nFracPredhi95), alpha = 0.5, fill = "darkgray") +
        geom_ribbon(aes(ymin = nFracPredlo80, ymax = nFracPredhi80), alpha = 1, fill = "darkgray") +
        geom_line(aes(y = nFracPred), color = "blue") +
        geom_line(linewidth = 1) +
        geom_point(size = 3) +
        labs(title = "Number of Hip Fractures", subtitle = "Compared to an ARIMA model using data from 2008 to 2015",
              caption = "Source: SIH-SUS")
        
p
download_this(p, output_extension = "svg")
```

Contrary to expectations based on the 2008-2015 trend, the number of fractures in the latter period exceeded predictions. A potential contributing factor to this unexpected surge could be the observed stagnation in the growth of densitometry exams during the same timeframe.

This plateau in densitometry testing rates suggests a potential unmet need in the diagnosis and early detection of osteoporosis—a critical risk factor for hip fractures. Consequently, the lack of preventative measures such as densitometry tests could be exacerbating the incidence of hip fractures.

It's important to note, however, that while the data indicates a correlation, further investigation would be required to establish a causal relationship between the stagnation in densitometry exams and the increased incidence of hip fractures.

Table for the plot above:

```{r}
#| label: hip-fractures-forecast
fracano %>% 
  mutate('80% Confience Interval' = if_else(is.na(nFracPred), NA_character_, paste(round(nFracPredlo80,0), " - ", round(nFracPredhi80,0))),
         '95% Confience Interval' = if_else(is.na(nFracPred), NA_character_, paste(round(nFracPredlo95,0), " - ", round(nFracPredhi95,0)))) %>% 
    rename(Ano = ano, 
         'Number of Fractures' = nFrac,
         'Predicted Number of Fractures' = nFracPred) %>% 
  select(-nFracPredlo80,-nFracPredhi80,-nFracPredlo95,-nFracPredhi95)
```

```{r}
#| label: hip-fractures-forecast-year

fracano <- RD %>% 
  filter(UNICO & CD_CID %in% c("S720","S721","S722")) %>% 
  filter(IDADE >= 50) %>% 
  filter(ano >= 2008) %>% 
  group_by(ano) %>% 
  summarise(nFrac = n_distinct(N_AIH)) %>% 
  left_join(popSUSBR %>%  mutate(idade = word(FaixaEtr)) %>%  filter (idade >= 50)%>% group_by(ano) %>% summarise(PopSUS = sum(PopSUS), .groups = "keep"), by = c("ano")) %>% 
  mutate(IndFrac = nFrac*100000/PopSUS)


model <- auto.arima(fracano$IndFrac[1:8])
forecast <- forecast(model)
forecast<- as.data.frame(forecast)

fracano$nFracPred <- NA
fracano$nFracPredlo80<- NA
fracano$nFracPredlo95<- NA
fracano$nFracPredhi80<- NA
fracano$nFracPredhi95<- NA


fracano$nFracPred[8] <- fracano$IndFrac[8]
fracano$nFracPredlo80[8] <- fracano$IndFrac[8]
fracano$nFracPredlo95[8] <- fracano$IndFrac[8]
fracano$nFracPredhi80[8] <- fracano$IndFrac[8]
fracano$nFracPredhi95[8] <- fracano$IndFrac[8]
fracano$nFracPred[9:15] <- forecast$`Point Forecast`[1:7]
fracano$nFracPredlo80[9:15] <- forecast$`Lo 80`[1:7]
fracano$nFracPredlo95[9:15] <- forecast$`Lo 95`[1:7]
fracano$nFracPredhi80[9:15] <- forecast$`Hi 80`[1:7]
fracano$nFracPredhi95[9:15] <- forecast$`Hi 95`[1:7]

p <- fracano %>% 
  ggplot(aes(x = ano, y = IndFrac)) +
        theme_bw() +
        scale_y_continuous(limits = c(50,130), labels = comma) +
        labs(x = "Year", y = "Hip Fractures Incidence Rate (per 100.000)") +
        geom_ribbon(aes(ymin = nFracPredlo95, ymax = nFracPredhi95), alpha = 0.5, fill = "darkgray") +
        geom_ribbon(aes(ymin = nFracPredlo80, ymax = nFracPredhi80), alpha = 1, fill = "darkgray") +
        geom_line(aes(y = nFracPred), color = "blue") +
        geom_line(linewidth = 1) +
        geom_point(size = 3) +
        labs(title = "Hip Fractures Incidence Rate", subtitle = "Compared to an ARIMA model using data from 2008 to 2015",
              caption = "Source: SIH-SUS")
        
p
download_this(p, output_extension = "svg")
```

When considering rates, the trend has a null growth which means that the growth in absolute numbers comes mainly from population growth. except in the period from 2015 to 2019 in which the growth completely detach from the trend. So it can be confirmed that the exceeding rates seen after 2015 is not caused by population growth, but on actual fracture rates increase.

```{r}
#| label: hip-fractures-forecast-rate-data
fracano %>% 
  mutate('80% Confience Interval' = if_else(is.na(nFracPred), NA_character_, paste(round(nFracPredlo80,0), " - ", round(nFracPredhi80,0))),
         '95% Confience Interval' = if_else(is.na(nFracPred), NA_character_, paste(round(nFracPredlo95,0), " - ", round(nFracPredhi95,0)))) %>% 
    rename(Ano = ano, 
         'Number of Fractures' = nFrac,
         'Predicted Number of Fractures' = nFracPred) %>% 
  select(-nFracPredlo80,-nFracPredhi80,-nFracPredlo95,-nFracPredhi95)
```

## Gender Disparities in Osteoporosis and Densitometries

```{r}
#| label: hip-fractures-bygender-number

df_fracrate <- RD %>% 
  filter(UNICO & CD_CID %in% c("S720","S721","S722")) %>% 
  filter(ano >= 2008) %>% 
  filter(IDADE >= 50) %>% 
  group_by(ano, SEXO) %>% 
  summarise(nFrac = n_distinct(N_AIH), .groups = "keep") %>% 
  left_join(popSUSBR %>% group_by(ano, Sexo) %>% filter(word(FaixaEtr,1) >= 50) %>% summarise(PopSUS = sum(PopSUS), .groups = "keep"), by = c("ano", "SEXO" = "Sexo")) %>% 
  mutate(IndFrac = nFrac*100000/PopSUS)

p <- df_fracrate %>% 
  ggplot(aes(x = ano, y = nFrac, color = SEXO)) +
        geom_line(linewidth = 1) +
        geom_point(size = 3)+
        theme_bw() +
        scale_y_continuous(limits = c(0,40000), labels = comma) +
        labs(x = "Year", y = "Number of hip fractures", title = "Hip fracture number by gender", color = "Gender")
p
download_this(p, output_extension = "svg")


```

The plot above is the same, but split by sex. Although it look like the fracture in women has grown faster, this is not the case since the numbers are proportional through all the period as can be seen in the last plot (fracture percentage in men).

```{r}
#| label: hip-fractures-bygender-rate
P <- df_fracrate %>% 
  ggplot(aes(x = ano, y = IndFrac, color = SEXO)) +
        geom_line(linewidth = 1) +
        geom_point(size = 3)+
        theme_bw() +
        scale_y_continuous(limits = c(0,150), labels = comma) +
        labs(x = "Year", y = "Hip fracture rates by 100.000 population", title = "Hip fracture rate by gender", color = "Gender")

P
download_this(p, output_extension = "svg")
```

```{r}

 df_fracrate %>% 
    pivot_longer(cols = c(nFrac, PopSUS, IndFrac), names_to = "variable", values_to = "value") %>% 
    pivot_wider(names_from = ano, values_from = value) %>% 
    mutate(across(where(is.numeric), ~prettyNum(.x, big.mark = ",", scientific = FALSE, trim = TRUE, )))
```

The same plot in rates.

### Hip Fracture Rates in men and women

```{r}
#| label: hip-fractures-bygender-rates-plot
FracTotRateSexo <- RD %>% filter (year(DT_INTER) == 2022) %>% 
  #filter(IDADE >= 50) %>% 
  filter(UNICO & CD_CID %in% c("S720","S721","S722")) %>% 
  group_by(FaixaEtr, SEXO) %>%
  summarize(n = n_distinct(N_AIH), .groups = "keep") %>% 
  left_join(popSUSBR %>% filter(ano == 2022) %>% group_by(FaixaEtr, Sexo) %>% summarise(PopSUS = sum(PopSUS), .groups = "keep"), by = c("FaixaEtr", "SEXO" = "Sexo")) %>%
  mutate(FracRate = n*100000/PopSUS)

levels(FracTotRateSexo$FaixaEtr) <- c("0 to 4 years", "5 to 9 years", "10 to 14 years", "15 to 19 years", "20 to 24 years", "25 to 29 years", "30 to 34 years", "35 to 39 years", "40 to 44 years", "45 to 49 years", "50 to 54 years", "55 to 59 years", "60 to 64 years", "65 to 69 years", "70 to 74 years", "75 to 79 years", "80 years or more")

FracRate <- ggplot(FracTotRateSexo, aes(x = FaixaEtr, y = if_else(SEXO == "Male", -FracRate,FracRate), fill = SEXO)) + 
  geom_bar(data = subset (FracTotRateSexo, SEXO == "Female"), stat = "identity") + 
  geom_bar(data = subset (FracTotRateSexo, SEXO == "Male"), stat = "identity") + 
  scale_y_continuous(labels = abs, limits = max(FracTotRateSexo$FracRate) * c(-1,1))  + 
  coord_flip() + 
  scale_fill_brewer(palette = "Set1") + 
  theme_bw() +
  labs(x = "Age Group", y = "Hip Fracture Rate per 100,000 population", title = "Hip Fracture Rate by Age Group and Sex. Brazil 2022",
        caption = "Source: IBGE, ANS, SIH-SUS (2022)", fill = "Gender")

FracRate

download_this(FracRate, output_extension = "svg")

```

The provided graph illustrates the incidence of hip fractures in men and women across various age groups. Notably, the rates of hip fracture in men appear to closely mirror those in women, both in terms of absolute numbers and the pattern of increase with age.

Contrary to common assumptions, our data reveals that the incidence of hip fractures in men aged 60 to 64 is actually slightly higher than that in women of the same age group. This similarity in fracture rates across genders suggests that the age at which individuals begin densitometry testing should be the same for both men and women.

Furthermore, our findings indicate that hip fractures in men should often be classified as fragility fractures rather than impact fractures. This is supported by the observation that hip fractures are relatively rare in men under 50, an age range where one might expect a higher incidence of impact-related fractures due to physical activity. The fact that the incidence of fractures increases dramatically in men over 50 – an age group unlikely to experience a significant increase in high-impact activities – suggests that declining bone density, rather than increased impact, is likely to be the primary cause.

In conclusion, these data suggest a need to revisit assumptions about gender differences in bone health and fracture risk, and to consider more carefully the role of bone density in fracture risk among older men.

```{r}
#| label: hip-fractures-bygender-median-age
RD %>% 
  filter(UNICO & CD_CID %in% c("S720","S721","S722")) %>% 
  group_by(SEXO) %>%
  summarise(median_age = median(IDADE),
            perc25 = quantile(IDADE, probs = 0.25),
            perc75 = quantile(IDADE, probs = 0.75))
```

```{r}
#| label: hip-fractures-bygender-rate-age-groups
FracTotRateSexo %>% 
  select(FaixaEtr, SEXO, FracRate) %>% 
  pivot_wider(names_from = SEXO, values_from = FracRate) %>%
  mutate(FaixaEtr = fct_relevel(FaixaEtr, "0 to 4 years", "5 to 9 years", "10 to 14 years", "15 to 19 years", "20 to 24 years", "25 to 29 years", "30 to 34 years", "35 to 39 years", "40 to 44 years", "45 to 49 years", "50 to 54 years", "55 to 59 years", "60 to 64 years", "65 to 69 years", "70 to 74 years", "75 to 79 years", "80 years or more"))
```

Fracture rate in people over 70 years old

```{r}
#| label: hip-fractures-bygender-rate-table
#Fracture ratio for men and women over 70 years old grouped by gender
FracTotRateSexo %>% 
  filter(FaixaEtr %in% c("70 to 74 years", "75 to 79 years", "80 years or more")) %>% 
  group_by(SEXO) %>%
  summarize(n = sum(n), PopSUS = sum(PopSUS), .groups = "keep") %>%
  mutate(FracRate = n*100000/PopSUS)

```

```{r}
#| label: hip-fractures-odds_r-table
FracRateOR <- FracTotRateSexo %>% 
  group_by(SEXO) %>%
  arrange(SEXO, FaixaEtr) %>%
  #add cummulative sum of n and PopSUS
  mutate(n_cum = cumsum(n),
         PopSUS_cum = cumsum(PopSUS)) %>%
  #get the cumsum of n and PopSUS for the previous age group
  mutate(n_cum_lag = lag(n_cum),
         PopSUS_cum_lag = lag(PopSUS_cum),
         n_lag = lag(n),
         PopSUS_lag = lag(PopSUS)) %>%
  filter(FaixaEtr != "0 to 4 years") 


#loop function calculating OR and p value and adding to different columns for each line of FracRateOR using n, PopSUS, n_cum_lag, PopSUS_cum_lag

FracRateOR$OR_tot <- NA
FracRateOR$ORL_tot <- NA
FracRateOR$ORH_tot <- NA
FracRateOR$p_tot <- NA
FracRateOR$OR_ant <- NA
FracRateOR$ORL_ant <- NA
FracRateOR$ORH_ant <- NA
FracRateOR$p_ant <- NA


for (i in 1:nrow(FracRateOR)) {
  ORtable <- matrix(c(FracRateOR$n[i],FracRateOR$PopSUS[i]-FracRateOR$n[i],FracRateOR$n_cum_lag[i],FracRateOR$PopSUS_cum_lag[i]-FracRateOR$n_cum_lag[i]),nrow = 2, ncol = 2)
  OR <- oddsratio.wald(ORtable)
  FracRateOR$OR_tot[i] <- OR$measure[2,1]
  FracRateOR$ORL_tot[i] <- OR$measure[2,2]
  FracRateOR$ORH_tot[i] <- OR$measure[2,3]
  FracRateOR$p_tot[i] <- OR$p.value[2]
  
  ORtable <- matrix(c(FracRateOR$n[i],FracRateOR$PopSUS[i]-FracRateOR$n[i],FracRateOR$n_lag[i],FracRateOR$PopSUS_lag[i]-FracRateOR$n_lag[i]),nrow = 2, ncol = 2)
  OR <- oddsratio.wald(ORtable)
  FracRateOR$OR_ant[i] <- OR$measure[2,1]
  FracRateOR$ORL_ant[i] <- OR$measure[2,2]
  FracRateOR$ORH_ant[i] <- OR$measure[2,3]
  FracRateOR$p_ant[i] <- OR$p.value[2]


}
FracRateOR


```

```{r}
#| label: hip-fractures-odds_r-plot
P <- FracRateOR %>% 
  ggplot(aes(x = FaixaEtr, y = OR_ant, ymin = ORL_ant, ymax = ORH_ant, color = SEXO, label = format(OR_ant, nsmall = 2 , digits = 2))) +
  geom_point() +
  geom_errorbar(width = 0.2) +
  theme_bw() +
  #add a horizontal line at y = 1 to see where OR = 1 
  geom_hline(yintercept = 1, linetype = "dashed", alpha = 0.5) +
  #add labels to the points
  geom_text(position = position_dodge(width = 0.1), hjust = -0.15, size = 3, show.legend = FALSE) +
  #x axis labels in 45 degres
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  #yaxis in log scale
  scale_y_log10() +
  labs(y = "Odds Ratio in relation to the previous age group",
        x = "Age Range",
        title = "Odds Ratio of hip fractures in men and women",
        subtitle = "Brazil, 2022", color = "Sex")

P

download_this(P, output_extension = "svg")

```

```{r}
#| label: hip-fractures-odds_r-plot-previous-age
P <- FracRateOR %>% 
  ggplot(aes(x = FaixaEtr, y = OR_tot, ymin = ORL_tot, ymax = ORH_tot, color = SEXO, label = format(OR_tot, nsmall = 2 , digits = 2))) +
  geom_point() +
  geom_errorbar( width = 0.10) +
  theme_bw() +
  #add a horizontal line at y = 1 to see where OR = 1 
  geom_hline(yintercept = 1, linetype = "dashed", alpha = 0.5) +
  #add labels to the points
  geom_text(position = position_dodge(width = 0.1), hjust = -0.3, size = 4, show.legend = FALSE) +
  #x axis labels in 45 degres
  theme(axis.text.x = element_text(angle = 45, hjust = 1, size = 12)) +
  #yaxis in log scale
  scale_y_log10() +
  #extend the limits of the x axis the text is not cut
  scale_x_discrete(expand = c(0.05, 0.12)) +
  
  labs(y = "Odds Ratio in relation to previous ages",  
        x = "Age Range",
        title = "Odds Ratio of hip fractures in men and women",
        subtitle = "Brazil, 2022", color = "Sex")

P
  
download_this(P, output_extension = "svg")

```

### Densitometry Rates in men and women

```{r}
#| label: densitometry-gender-age-group-plot
DensTotRateSexo <- PA %>% filter (year(DATA) == 2022) %>% 
  filter(SEXO != "Undefined") %>% 
  group_by(FaixaEtr, SEXO) %>%
  summarize(n = sum(PA_QTDPRO), .groups = "keep") %>% 
  left_join(popSUSBR %>% filter(ano == 2022) %>% group_by(FaixaEtr, Sexo) %>% summarise(PopSUS = sum(PopSUS), .groups = "keep"), by = c("FaixaEtr", "SEXO" = "Sexo")) %>%
  mutate(DensRate = n*100000/PopSUS)

levels(DensTotRateSexo$FaixaEtr) <- c("0 to 4 years", "5 to 9 years", "10 to 14 years", "15 to 19 years", "20 to 24 years", "25 to 29 years", "30 to 34 years", "35 to 39 years", "40 to 44 years", "45 to 49 years", "50 to 54 years", "55 to 59 years", "60 to 64 years", "65 to 69 years", "70 to 74 years", "75 to 79 years", "80 years or more")

DensRate <- ggplot(DensTotRateSexo, aes(x = FaixaEtr, y = if_else(SEXO == "Male", -DensRate,DensRate), fill = SEXO)) + 
  geom_bar(data = subset (DensTotRateSexo, SEXO == "Female"), stat = "identity") + 
  geom_bar(data = subset (DensTotRateSexo, SEXO == "Male"), stat = "identity") + 
  scale_y_continuous(labels = abs, limits = max(DensTotRateSexo$DensRate) * c(-1,1))  + 
  coord_flip() + 
  scale_fill_brewer(palette = "Set1") + 
  theme_bw() +
  labs(x = "Age Group", y = "Bone Densitometry Rate per 100,000 inhabitants", title = "Bone Densitometry Rate by Age Group and Sex",
        caption = "Source: IBGE, ANS, SIA-SUS (2022)", fill = "Sex")

DensRate

download_this(DensRate, output_extension = "svg")
```

The plot displayed above represents the frequency of bone density tests conducted in Brazil in 2022. A striking observation from the data is that men undergo these densitometry exams at significantly lower rates compared to women.

Furthermore, when these figures are compared with the rates of hip fractures, an apparent gender disparity in bone density examinations emerges that does not align with the fracture risk profile. Despite similar risks for hip fractures across genders, as seen in our previous analysis, men appear to have far fewer densitometry tests.

This discrepancy suggests that the current approach to bone health monitoring may not be adequately addressing the needs of men, and invites a closer look at the gender dynamics of healthcare accessibility and utilization.

**Densitometry rate in men and women over 70 years old**
```{r}
#| label: densitometry-gender-over70-table
# Densitometry rate in men and women over 70 years old
DensTotRateSexo %>% 
  filter(FaixaEtr %in% c("60 to 64 years", "65 to 69 years","70 to 74 years", "75 to 79 years", "80 years or more")) %>% 
  group_by(SEXO) %>%
  summarize(n = sum(n), PopSUS = sum(PopSUS), .groups = "keep") %>%
  mutate(DensRate = n*100000/PopSUS)
```

### Rates over time

```{r}
#| label: dens-frac-rates-disparity

dens_rate_raw <- PA %>% 
  filter(SEXO != "Undefined") %>% 
  group_by(FaixaEtr, SEXO, ano) %>%
  summarize(n = sum(PA_QTDPRO), .groups = "keep") %>% 
  left_join(popSUSBR %>% group_by(FaixaEtr, Sexo, ano) %>% summarise(PopSUS = sum(PopSUS), .groups = "keep"), by = c("FaixaEtr", "SEXO" = "Sexo", "ano")) %>% 
  filter(as.numeric(word(FaixaEtr))>=50) %>% 
  group_by(SEXO, ano) %>% 
  summarise(n = sum(n),
            PopSUS = sum(PopSUS), .groups = "keep") %>% 
  mutate(DensRate = n*100000/PopSUS)


DensRateSex <- dens_rate_raw %>% 
  pivot_wider(id_cols = ano, names_from = SEXO, values_from = DensRate) %>% 
  mutate(MaleRateDens = Male/Female)

FracRateSex <-RD %>% 
  filter(UNICO & CD_CID %in% c("S720","S721","S722")) %>% 
  filter(SEXO != "Undefined") %>% 
  filter(IDADE >= 50) %>% 
  group_by(FaixaEtr, SEXO, ano) %>%
  summarize(n = n_distinct(N_AIH), .groups = "keep") %>% 
  left_join(popSUSBR %>% group_by(FaixaEtr, Sexo, ano) %>% filter(word(FaixaEtr,1) >= 50) %>% summarise(PopSUS = sum(PopSUS), .groups = "keep"), by = c("FaixaEtr", "SEXO" = "Sexo", "ano")) %>% 
  filter(as.numeric(word(FaixaEtr))>=50) %>% 
  group_by(SEXO, ano) %>% 
  summarise(n = sum(n),
            PopSUS = sum(PopSUS), .groups = "keep") %>% 
  mutate(FracRate = n*100000/PopSUS) %>% 
  pivot_wider(id_cols = ano, names_from = SEXO, values_from = FracRate) %>% 
  mutate(MaleRateFrac = Male/Female)


p <- FracRateSex %>% 
  select(ano, MaleRateFrac) %>% 
  left_join(DensRateSex %>% select(ano, MaleRateDens), by = "ano") %>% 
  filter(ano >=2008) %>% 
  ggplot(aes(x = ano)) +
  geom_line(aes(y = MaleRateFrac, color = "Hip Fractures"), linewidth = 1) +
  geom_point(aes(y = MaleRateFrac, color = "Hip Fractures"), size = 2) +
  geom_line(aes(y = MaleRateDens, color = "Bone density scans"), linewidth = 1) +
  geom_point(aes(y = MaleRateDens, color = "Bone density scans"), size = 2) +
  scale_y_continuous(limits = c(0,0.65)) +
  labs(x = "Year", y = "Male to Female Rate Ratio", color = "", title = "Male to Female Hip Fractures and Bone Density Scans Rate Ratios",
        subtitle = "Considering fractures only in people over 50 years of age",
        caption = "Source: SIA-SUS, SIH-SUS, IBGE, ANS (2008-2022)") +
  theme_classic()
p
  
download_this(p, output_extension = "svg")
```

\*The interpretation of a rate ratio of 0.6 is that for every 100 hip fracture in women, there are 60 in men.

The graph presented above illustrates a significant disparity between the incidence of hip fractures and the rate of preventative bone density testing in men. While the rate ratio of hip fractures occurring in men remains relatively stable at around 60, the ratio of men undergoing bone density scans is markedly lower, hovering below 10%.

Despite a modest upward trend over time in the rate of bone density scans among men, the gap between the observed fracture risk and preventative measures remains substantial. This discrepancy, when viewed in light of our previous discussion on the similarities in fracture risks between genders, underscores a potential under-utilization of bone health evaluations in men.

Thus, these findings highlight an essential need for improved bone health strategies that better align preventative measures, like bone density tests, with the actual risks faced by different population groups. This may involve increasing awareness about the risk of fractures in men and emphasizing the importance of preventative measures like bone density scans.

**Densitometry numbers >50**
```{r}
#| label: dens-frac-rates-disparity_table
dens_rate_raw %>% 
  mutate(SEXO = droplevels(SEXO)) %>%
  pivot_longer(cols = c(n, PopSUS, DensRate), names_to = "variable", values_to = "value") %>% 
  pivot_wider(names_from = ano, values_from = value) %>% 
  mutate(across(where(is.numeric), ~prettyNum(.x, big.mark = ",", scientific = FALSE, trim = TRUE, )))
```

**Densitometry Numbers for Male >70 and Female >65**
```{r}
#| label: dens-numbers_segmented
dens_rate_raw_segmented <- PA %>% 
  filter(SEXO != "Undefined") %>% 
  group_by(FaixaEtr, SEXO, ano) %>%
  summarize(n = sum(PA_QTDPRO), .groups = "keep") %>% 
  left_join(popSUSBR %>% group_by(FaixaEtr, Sexo, ano) %>% summarise(PopSUS = sum(PopSUS), .groups = "keep"), by = c("FaixaEtr", "SEXO" = "Sexo", "ano")) %>% 
  filter((SEXO == "Female" & as.numeric(word(FaixaEtr))>=65) | (SEXO == "Male" & as.numeric(word(FaixaEtr))>=70)) %>% 
  group_by(SEXO, ano) %>% 
  summarise(n = sum(n),
            PopSUS = sum(PopSUS), .groups = "keep") %>% 
  mutate(DensRate = n*100000/PopSUS)

dens_rate_raw_segmented %>% 
  mutate(SEXO = droplevels(SEXO)) %>%
  pivot_longer(cols = c(n, PopSUS, DensRate), names_to = "variable", values_to = "value") %>% 
  pivot_wider(names_from = ano, values_from = value) %>% 
  mutate(across(where(is.numeric), ~prettyNum(.x, big.mark = ",", scientific = FALSE, trim = TRUE, )))
``` 

## Bone Density Scanners

```{r}
#| label: dx-scanners
p <- EQ %>% 
  group_by(year = year(DATA), SUS) %>% 
  summarise(QT = sum(QT),
            QT_USO = sum(QT_USO), .groups = "keep") %>% 
  ggplot(aes(x = year, y = QT_USO, color = SUS)) +
  geom_line() + 
  geom_point() +
  scale_y_continuous(limits = c(0,20000)) +
  theme_classic()+
  labs(title = "Number of Densitometry Scans", x = "Year", y = "Number of Densitometry Scans", color = "Use in SUS")
p
download_this(p, output_extension = "svg")

```

```{r}
#| label: dx-scanners-rate
dx_rate <- EQ %>% 
  group_by(year = year(DATA), SUS) %>% 
  summarise(QT = sum(QT),
            QT_USO = sum(QT_USO), .groups = "keep") %>% 
  left_join(popSUSBR %>% group_by(ano) %>% filter(word(FaixaEtr,1) >= 50) %>%
              summarise(SUS = sum(PopSUS), `Non-SUS` = sum(PopANS), .groups = "keep") %>% 
              pivot_longer(cols = c("SUS", "Non-SUS"), names_to = "system", values_to = "Pop")
              , by = c("year"= "ano", "SUS" = "system")) %>% 
  mutate(densrate = QT_USO*100000/Pop)



p <- dx_rate %>% 
  ggplot(aes(x = year, y = densrate, color = SUS)) +
  geom_line() + 
  geom_point() +
  scale_y_continuous(limits = c(0,170)) +
  theme_classic()+
  labs(title = "Densitometry Scans per Population", x = "Year", y = "Densitometry Scans per 100.000 people aged 50+", color = "Use in SUS")
p

download_this(p, output_extension = "svg")
```

```{r}
#| label: dx-scanners-table
rateSUS <- dx_rate %>%  filter(year == 2022) %>% ungroup() %>% select (SUS, densrate) %>% pivot_wider(names_from = SUS, values_from = densrate) %>% 
  mutate (n = SUS/`Non-SUS`) %>% pull(n)

  
dx_rate
  
```

In 2022, there was {round(1/rateSUS,1)} times more Densitometry Scans in private healthcare in comparison to SUS. The Densitometry Scans rate is stable in SUS since 2015.

## Table with procedures codes
Number of hip fractures in SUS:
```{r}
#| label: raw-table-hip-frac
#Generate table with procedures codes used in the Hip Fracture analysis RD
#Filtering only the procedures used in the analysis


#Annual Fracture rate by gender in the whole time period in RD with the SUS population in the same period 
RD %>% 
  filter(UNICO & CD_CID %in% c("S720","S721","S722")) %>% 
  filter(IDADE >= 50) %>% 
  group_by(ano, SEXO) %>% 
  summarise(n = n(), .groups = "keep") %>% 
  pivot_wider(names_from = SEXO, values_from = n)
```

Median of age in hip fracture hospitalisations:
```{r}
#| label: raw-table-hip-frac-age
#Median of age in hip fracture hospitalisations
RD %>% 
  filter(UNICO & CD_CID %in% c("S720","S721","S722")) %>% 
  filter(IDADE >= 50) %>% 
  group_by(SEXO) %>% 
  summarise(median_age = median(IDADE), .groups = "keep", 
            percentile25 = quantile(IDADE, 0.25),
            percentile75 = quantile(IDADE, 0.75))


```

# Adjusted Rates

```{r}
#| label: adjusted-frac-rate
FracStdRate <- RD %>% 
  filter(UNICO & CD_CID %in% c("S720","S721","S722")) %>% 
  filter(ano >= 2008) %>% 
  filter(IDADE >= 50) %>% 
  group_by(ano, SEXO, FaixaEtr) %>% 
  summarise(nFrac = n_distinct(N_AIH), .groups = "keep") %>% 
  left_join(popSUSBR %>% group_by(ano, Sexo, FaixaEtr) %>% filter(word(FaixaEtr,1) >= 50) %>% summarise(PopSUS = sum(PopSUS), .groups = "keep"), by = c("ano", "SEXO" = "Sexo", "FaixaEtr")) %>% 
  group_by(ano, FaixaEtr) %>%
  summarise(PopSUS = sum(PopSUS),
            nFrac = sum(nFrac), .groups = "keep") %>%
  left_join(std_WHO, by = "FaixaEtr") %>%
  mutate(act_rate = nFrac/PopSUS,
         WHO_n = act_rate * WHO20002025) %>%
  group_by(ano) %>%
  summarise(PopSUS = sum(PopSUS),
            nFrac = sum(nFrac),
            WHO20002025 = sum(WHO20002025),
            WHO_n = sum(WHO_n)) %>% 
  mutate(std_rate = WHO_n*100000/WHO20002025,
         act_rate = nFrac*100000/PopSUS)

FracStdRate %>%
  ggplot(aes(x = ano)) +
  geom_line(aes(y = std_rate, color = "Standardised Rate"), linewidth = 1) + 
  geom_line(aes(y = act_rate, color = "Actual Rate"), linewidth = 1) +
  theme(axis.text.x = element_text(angle = 20, hjust = 1)) +
labs(caption = "Source: SIH-SUS, IBGE, ANS",
     subtitle = "Considering Population over 50 years old", 
     title = "Standardised vs Actual Hip Fracture Rates - Brazil, 2008-2022",
     x = "Year", y = "Rate per 100,000", color = "Rate Type") +
theme_bw() +
scale_y_continuous(limits = c(0, NA))
```

```{r}
#| label: adjusted-frac-rate-table
#Create a piovt table of FracStdRate with ano as column
FracStdRate %>%
pivot_longer(col= -ano) %>%
pivot_wider(names_from = ano, values_from = value) %>% 
mutate(across(everything(), ~formatC(., format = "f", digits = 0)))

```

The above plot and table, describes the difference between the actual and standardised rates of hip fractures in Brazil from 2008 to 2022. The actual rate is the number of hip fractures per 100,000 people aged 50 and over in Brazil. The standardised rate is the number of hip fractures per 100,000 people aged 50 and over in Brazil, adjusted for the age distribution of the world population in 2000-2025. The actual rate is the same as the standardised rate in 2008, but it increases in the following years. The actual rate is `r format(FracStdRate %>% filter(ano == 2022) %>% pull(act_rate), format = "f", digits = 4)` hip fractures per 100,000 people aged 50 and over in Brazil in 2022. The standardised rate is `r format(FracStdRate %>% filter(ano == 2022) %>% pull(std_rate), format = "f", digits = 4)` hip fractures per 100,000 people aged 50 and over in Brazil in 2022. This means that the aging of population is only responsible for a small part of the increase in hip fractures in Brazil.

Of the `r format(((FracStdRate %>% filter(ano == 2022) %>% pull(act_rate)/FracStdRate %>% filter(ano == 2008) %>% pull(act_rate))-1)*100,digits = 3)`% growth in hip fractures rate from 2008 to 2022, only `r format(((FracStdRate %>% filter(ano == 2022) %>% pull(act_rate)/FracStdRate %>% filter(ano == 2008) %>% pull(act_rate))/(FracStdRate %>% filter(ano == 2022) %>% pull(std_rate)/FracStdRate %>% filter(ano == 2008) %>% pull(std_rate))-1)*100,digits = 3)`% of the growth can be explained by the aging of the population.
